requirements.txt
asgiref==3.6.0
Django==3.2.16
django-environ==0.9.0
djangorestframework==3.14.0
pytz==2022.7
sqlparse==0.4.3
